+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Tigris Changelog
 
 <!---------------------------------->
+<a name="v0.2.0"></a>
+### v0.2.0
+- Added learning using evolution
+- Added `tigris::Environment`
+- Added `tigris::AI`
+- Added `tigris::Matrix`
+- Added game status Draw to tic-tac-toe
+
+
+<!---------------------------------->
 <a name="v0.1.0"></a>
 ### v0.1.0
 - Initial setup